ENH: Add apptainer functionality
Adds a setting in user_config that runs the segmentations with apptainer/singularity instead of docker. This is generally better because it does not require sudo acccess and also allows the pipeline to be run on compute cluster where that may be an issue. There's also a small hotfix to how a few scripts are called.